requirements.txt
streamlit==1.46.1
tensorflow==2.19.0
numpy==2.0.2
Pillow==11.3.0